added nav and side bars to the userDetails page as well as other changes requsted
