+++ package.json
@@ -28,8 +28,7 @@
     "react-hook-form": "^7.51.5",
     "react-redux": "^9.1.0",
     "react-router-dom": "^6.21.3",
-    "react-toastify": "^10.0.5",
-    "redux-persist": "^6.0.0"
+    "react-toastify": "^10.0.5"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
